requirements.txt
africastalking==1.2.3
amqp==2.6.1
asgiref==3.3.1
attrs==20.3.0
awsebcli==3.19.4
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
blessed==1.18.0
boto3==1.17.54
botocore==1.20.54
cached-property==1.5.2
celery==4.4.2
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.7.1
dj-rest-auth==1.1.0
Django==3.1.7
django-allauth==0.43.0
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-environ==0.4.5
django-extensions==2.0.7
django-filter==2.4.0
django-guardian==2.3.0
django-livereload-server==0.3.2
django-phonenumber-field==5.1.0
django-rest-auth==0.9.5
django-storages==1.11.1
django-timezone-field==4.1.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==6.6.2
future==0.16.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
ipython==7.21.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
kombu==4.6.11
Markdown==3.0.1
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
paramiko==2.7.2
parso==0.8.1
pathspec==0.5.9
pexpect==4.8.0
phonenumbers==8.12.20
pickleshare==0.7.5
prompt-toolkit==3.0.17
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
PyJWT==1.7.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.1
python3-openid==3.2.0
pytz==2018.5
PyYAML==5.3.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.4.0
schema==0.7.4
semantic-version==2.8.5
six==1.14.0
soupsieve==2.2.1
sqlparse==0.4.1
swagger-spec-validator==2.7.3
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.3
tornado==6.1
traitlets==5.0.5
typing==3.6.4
uritemplate==3.0.1
urllib3==1.25.11
vine==1.3.0
wcwidth==0.1.9
websocket-client==0.58.0
